+++ backend/src/api/sys_usage.rs
@@ -34,7 +34,7 @@ pub fn exec_system_usage(app_state: &Arc<AppState>) -> tokio::task::JoinHandle<(
 
                 state.event_manager.send_system_info(info);
             }
-            tokio::time::sleep(std::time::Duration::from_millis(1000)).await;
+            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
         }
     })
 }